Separate image side to component in authentication pages and fix height for app layout
