+++ src/App.module.scss
@@ -1,10 +1,20 @@
 @import "./shared/theme";
 
 .app {
+  height: 100vh;
   font-size: $NORMAL_FONT_SIZE;
+  display: flex;
+  flex-direction: column;
 
   &::-webkit-scrollbar-track {
     -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
     background-color: #f5f5f5;
+  }
+}
+
+.content {
+  flex-grow: 1;
+  & > div {
+    height: 100%;
   }
 }
