+++ src/Components/Authentication/SignUpConfirmation/SignUpConfirmation.module.scss
@@ -5,6 +5,7 @@
     color: $primary900;
     font-weight: 500;
     font-size: $NORMAL_FONT_SIZE;
+    height: 100%;
 }
 
 .rightSide {
