+++ src/shared/_break-point.scss
@@ -1,5 +1,5 @@
-$bp-xl: 1920px;
-$bp-lg: 1200px; 
-$bp-md: 992px;
-$bp-sm: 768px;
-$bp-xs: 576px; 
+$bp-xl: 1535px;
+$bp-lg: 1199px; 
+$bp-md: 899px;
+$bp-sm: 599px;
+$bp-xs: 0px;
