+++ src/shared/components/CustomPattern/CustomPattern.module.scss
@@ -1,12 +1,10 @@
 .leftPattern {
     position: absolute;
-    width: 50% !important;
     top: -3vh;
     left: 0vw;
 }
 .rightPattern {
     position: absolute;
-    width: 50% !important;
     top: -3vh;
     right: 0vw;
 }
